feat : can search for titles, authors, or keywords
The search feature is working properly. The value in the search feature will be set to the query parameter with the `q` key.

Setting the query parameter occurs when the form is submitted and runs the handleSubmit.

When the value in the search field is `empty string`, it will return immediately (not setting query parameters). Otherwise, it will set the query parameter and change the pathname to `/search` using the useRouter hook's `push` method.

Then create a `useEffect` hook which is useful when the Header component is rendered and will check the pathname (property of the `useRouter` hook) that URL `/search`. If there is it will automatically call the `push` method of the `useRouter` hook and set the query parameter with the key `q` while specifying the pathname to `/search`.

The values ​​in the search fields and functions to set or change their values ​​are stored in the `searchValue` and `setSearchValue` variables which are destructuring from the `useState` hook.